+++ lib/src/features/profile_other_user/widget/tab_view_profile_user.dart
@@ -6,10 +6,7 @@ import 'package:myapp/src/features/profile_other_user/logic/favorite_bloc.dart';
 import 'package:myapp/src/features/profile_other_user/logic/favorite_state.dart';
 import 'package:myapp/src/features/profile_other_user/logic/host_bloc.dart';
 import 'package:myapp/src/features/profile_other_user/logic/host_state.dart';
-import 'package:myapp/src/features/search/widget/event_search_widget.dart';
-import 'package:myapp/src/network/model/common/pagination/pagination.dart';
-import 'package:myapp/src/network/model/event/event.dart';
-import 'package:myapp/widgets/state/state_pagination_widget.dart';
+import 'package:myapp/src/features/profile_other_user/widget/list_data_paination_event.dart';
 
 class TabViewProfileUser extends StatelessWidget {
   const TabViewProfileUser({
@@ -47,45 +44,5 @@ class TabViewProfileUser extends StatelessWidget {
         }),
       ),
     ]);
-  }
-}
-
-// ignore: must_be_immutable
-class ListDataPaginationEvent<T> extends StatelessWidget {
-  ListDataPaginationEvent({
-    super.key,
-    required this.data,
-    required this.getData,
-  });
-
-  MPagination<T> data;
-  Function() getData;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Expanded(
-          child: ListView.builder(
-            itemCount: data.data.length + 1,
-            itemBuilder: ((context, index) {
-              if (index == data.data.length) {
-                return Container(
-                  margin: const EdgeInsets.symmetric(vertical: 20),
-                  alignment: Alignment.center,
-                  child: XStatePaginationWidget(
-                    page: data,
-                    loadMore: getData,
-                    autoLoad: true,
-                  ),
-                );
-              } else {
-                return EventSearchWidget(event: data.data[index] as MEvent);
-              }
-            }),
-          ),
-        ),
-      ],
-    );
   }
 }
